+++ apps/docs-app/src/pages/components/pages/collections_item_catalog.rs
@@ -1,0 +1,71 @@
+use crate::pages::components::ComponentDoc;
+
+pub(super) const ITEM_DOC: ComponentDoc = ComponentDoc {
+    name: "Item",
+    slug: "item",
+    group: "Collections",
+    page: super::collections_item_shadcn::item_primitives,
+};
+
+pub(super) const ITEM_GROUP_DOC: ComponentDoc = ComponentDoc {
+    name: "ItemGroup",
+    slug: "item-group",
+    group: "Collections",
+    page: super::collections_item_shadcn::item_primitives,
+};
+
+pub(super) const ITEM_SEPARATOR_DOC: ComponentDoc = ComponentDoc {
+    name: "ItemSeparator",
+    slug: "item-separator",
+    group: "Collections",
+    page: super::collections_item_shadcn::item_primitives,
+};
+
+pub(super) const ITEM_MEDIA_DOC: ComponentDoc = ComponentDoc {
+    name: "ItemMedia",
+    slug: "item-media",
+    group: "Collections",
+    page: super::collections_item_shadcn::item_primitives,
+};
+
+pub(super) const ITEM_CONTENT_DOC: ComponentDoc = ComponentDoc {
+    name: "ItemContent",
+    slug: "item-content",
+    group: "Collections",
+    page: super::collections_item_shadcn::item_primitives,
+};
+
+pub(super) const ITEM_TITLE_DOC: ComponentDoc = ComponentDoc {
+    name: "ItemTitle",
+    slug: "item-title",
+    group: "Collections",
+    page: super::collections_item_shadcn::item_primitives,
+};
+
+pub(super) const ITEM_DESCRIPTION_DOC: ComponentDoc = ComponentDoc {
+    name: "ItemDescription",
+    slug: "item-description",
+    group: "Collections",
+    page: super::collections_item_shadcn::item_primitives,
+};
+
+pub(super) const ITEM_ACTIONS_DOC: ComponentDoc = ComponentDoc {
+    name: "ItemActions",
+    slug: "item-actions",
+    group: "Collections",
+    page: super::collections_item_shadcn::item_primitives,
+};
+
+pub(super) const ITEM_HEADER_DOC: ComponentDoc = ComponentDoc {
+    name: "ItemHeader",
+    slug: "item-header",
+    group: "Collections",
+    page: super::collections_item_shadcn::item_primitives,
+};
+
+pub(super) const ITEM_FOOTER_DOC: ComponentDoc = ComponentDoc {
+    name: "ItemFooter",
+    slug: "item-footer",
+    group: "Collections",
+    page: super::collections_item_shadcn::item_primitives,
+};